Mirror the javascript toolchain layout: generic package graph in its own file, project graph in tier2
