+++ toolchains/go/src/lib.rs
@@ -4,7 +4,7 @@ pub mod go_sum;
 pub mod go_work;
 
 #[cfg(feature = "wasm")]
-mod project_graph;
+mod package_graph;
 #[cfg(feature = "wasm")]
 mod tier1;
 #[cfg(feature = "wasm")]
@@ -12,8 +12,6 @@ mod tier2;
 #[cfg(feature = "wasm")]
 mod tier3;
 
-#[cfg(feature = "wasm")]
-pub use project_graph::*;
 #[cfg(feature = "wasm")]
 pub use tier1::*;
 #[cfg(feature = "wasm")]
